Edit employee fix

+++ CarWorkshop.Infrastructure/DTO/EmployeeDTO.cs
@@ -13,7 +13,7 @@ namespace CarWorkshop.Infrastructure.DTO
         public DateTime EmploymentDate { get; set; }
         public string PhoneNumber { get; set; }
         public string EmailAddress { get; set; }
-        public int Salary { get; set; }
+        public decimal Salary { get; set; }
         public string Currency { get; set; }
         public string Position { get; set; }
         public string UserRole { get; set; }
